changed my libft

+++ ft_ls/libft/ft_itoa.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_itoa.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fremoor <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: fremoor <fremoor@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/21 08:29:17 by fremoor           #+#    #+#             */
-/*   Updated: 2019/06/18 09:17:11 by fremoor          ###   ########.fr       */
+/*   Updated: 2019/07/10 15:26:33 by fremoor          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "./includes/libft.h"
 
 /*
 ** n_len gets the length of the number as chars, e.g. 10 would be 2 and 234 = 3
 */
 
-static int	n_len(int n)
+static int	n_len(long long n)
 {
 	int i;
 
@@ -36,11 +36,11 @@ static int	n_len(int n)
 ** the last number and then divs it by 10 to move it up one digit
 */
 
-char		*ft_itoa(int n)
+char		*ft_itoa(long long n)
 {
-	int		neg;
-	int		len;
-	char	*str;
+	int			neg;
+	long long	len;
+	char		*str;
 
 	neg = (n < 0 ? 1 : 0);
 	len = n_len(n);
